requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
gradio==4.12.0
celery==5.3.4
redis==5.0.1
minio==7.1.17
python-magic-bin==0.4.14; sys_platform == 'win32'
pydub==0.25.1
soundfile==0.12.1
ffmpeg-python==0.2.0
faster-whisper==0.9.0
torch==2.1.0
torchaudio==2.1.0
pyannote.audio==3.1.1
aiofiles==23.2.1
python-magic-bin==0.4.14; sys_platform == 'win32'
